Bot-Captura-Produtos/requirements.txt
mysql-connector-python
selenium
flask